Refactor co/case in fun
