+++ lang/fun/src/syntax/mod.rs
@@ -1,6 +1,3 @@
-use printer::{tokens::COMMA, util::BracesExt, Alloc, Builder, DocAllocator, Print, PrintCfg};
-use terms::Clause;
-
 pub type Variable = String;
 pub type Covariable = String;
 pub type Name = String;
@@ -10,30 +7,3 @@ pub mod declarations;
 pub mod substitution;
 pub mod terms;
 pub mod types;
-
-pub fn print_cases<'a>(
-    cases: &'a [Clause<Name>],
-    cfg: &PrintCfg,
-    alloc: &'a Alloc<'a>,
-) -> Builder<'a> {
-    match cases.len() {
-        0 => alloc.space().braces_anno(),
-
-        1 => alloc
-            .line()
-            .append(cases[0].print(cfg, alloc))
-            .nest(cfg.indent)
-            .append(alloc.line())
-            .braces_anno()
-            .group(),
-        _ => {
-            let sep = alloc.text(COMMA).append(alloc.hardline());
-            alloc
-                .hardline()
-                .append(alloc.intersperse(cases.iter().map(|x| x.print(cfg, alloc)), sep.clone()))
-                .nest(cfg.indent)
-                .append(alloc.hardline())
-                .braces_anno()
-        }
-    }
-}
