--- lang/fun/src/syntax/terms/case.rs
@@ -1,331 +1,0 @@
-use std::rc::Rc;
-
-use codespan::Span;
-use derivative::Derivative;
-use printer::{
-    theme::ThemeExt,
-    tokens::{CASE, DOT},
-    DocAllocator, Print,
-};
-
-use crate::{
-    parser::util::ToMiette,
-    syntax::{
-        context::TypingContext,
-        print_cases,
-        types::{OptTyped, Ty},
-        Name,
-    },
-    typing::{check::Check, errors::Error, symbol_table::SymbolTable},
-};
-
-use super::{Clause, Term};
-
-#[derive(Derivative, Debug, Clone)]
-#[derivative(PartialEq, Eq)]
-pub struct Case {
-    #[derivative(PartialEq = "ignore")]
-    pub span: Span,
-    pub destructee: Rc<Term>,
-    pub cases: Vec<Clause<Name>>,
-    pub ty: Option<Ty>,
-}
-
-impl OptTyped for Case {
-    fn get_type(&self) -> Option<Ty> {
-        self.ty.clone()
-    }
-}
-
-impl Print for Case {
-    fn print<'a>(
-        &'a self,
-        cfg: &printer::PrintCfg,
-        alloc: &'a printer::Alloc<'a>,
-    ) -> printer::Builder<'a> {
-        self.destructee
-            .print(cfg, alloc)
-            .append(DOT)
-            .append(alloc.keyword(CASE))
-            .append(alloc.space())
-            .append(print_cases(&self.cases, cfg, alloc))
-    }
-}
-
-impl From<Case> for Term {
-    fn from(value: Case) -> Self {
-        Term::Case(value)
-    }
-}
-
-impl Check for Case {
-    fn check(
-        self,
-        symbol_table: &SymbolTable,
-        context: &TypingContext,
-        expected: &Ty,
-    ) -> Result<Self, Error> {
-        // Find out the type on which we pattern match by inspecting the first case.
-        // We throw an error for empty cases.
-        let (ty, mut expected_ctors) = match self.cases.first() {
-            Some(case) => symbol_table.lookup_ty_for_ctor(&self.span.to_miette(), &case.xtor)?,
-            None => {
-                return Err(Error::EmptyMatch {
-                    span: self.span.to_miette(),
-                })
-            }
-        };
-
-        // We check the "e" in "case e of {...}" against this type.
-        let destructee_checked = self.destructee.check(symbol_table, context, &ty)?;
-
-        let mut new_cases = vec![];
-        for case in self.cases {
-            if !expected_ctors.remove(&case.xtor) {
-                return Err(Error::UnexpectedCtorInCase {
-                    span: case.span.to_miette(),
-                    ctor: case.xtor.clone(),
-                });
-            }
-            match symbol_table.ctors.get(&case.xtor) {
-                Some(ctor_ctx) => {
-                    case.context.no_dups(case.xtor.clone())?;
-                    case.context.compare_to(ctor_ctx)?;
-
-                    let mut new_context = context.clone();
-                    new_context
-                        .bindings
-                        .append(&mut case.context.bindings.clone());
-
-                    let new_rhs = case.rhs.check(symbol_table, &new_context, expected)?;
-                    new_cases.push(Clause {
-                        rhs: new_rhs,
-                        ..case
-                    });
-                }
-                None => {
-                    return Err(Error::Undefined {
-                        span: case.span.to_miette(),
-                        name: case.xtor.clone(),
-                    })
-                }
-            }
-        }
-        if !expected_ctors.is_empty() {
-            return Err(Error::MissingCtorsInCase {
-                span: self.span.to_miette(),
-            });
-        }
-        Ok(Case {
-            destructee: destructee_checked,
-            cases: new_cases,
-            ty: Some(expected.clone()),
-            ..self
-        })
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use super::{Check, Term};
-    use crate::{
-        parser::fun,
-        syntax::context::TypingContext,
-        syntax::{
-            terms::{Case, Clause, Lit, Var},
-            types::Ty,
-        },
-        test_common::{symbol_table_list, symbol_table_tup},
-    };
-    use codespan::Span;
-    use printer::Print;
-    use std::rc::Rc;
-
-    #[test]
-    fn check_case_list() {
-        let mut ctx_case = TypingContext::default();
-        ctx_case.add_var("x", Ty::mk_i64());
-        ctx_case.add_var("xs", Ty::mk_decl("ListInt"));
-        let mut ctx = TypingContext::default();
-        ctx.add_var("x", Ty::mk_decl("ListInt"));
-        let symbol_table = symbol_table_list();
-        let result = Case {
-            span: Span::default(),
-            cases: vec![
-                Clause {
-                    span: Span::default(),
-                    xtor: "Nil".to_owned(),
-                    context: TypingContext::default(),
-                    rhs: Lit::mk(1).into(),
-                },
-                Clause {
-                    span: Span::default(),
-                    xtor: "Cons".to_owned(),
-                    context: ctx_case.clone(),
-                    rhs: Var::mk("x").into(),
-                },
-            ],
-            destructee: Rc::new(Var::mk("x").into()),
-            ty: None,
-        }
-        .check(&symbol_table, &ctx, &Ty::mk_i64())
-        .unwrap();
-        let expected = Case {
-            span: Span::default(),
-            cases: vec![
-                Clause {
-                    span: Span::default(),
-                    xtor: "Nil".to_owned(),
-                    context: TypingContext::default(),
-                    rhs: Lit::mk(1).into(),
-                },
-                Clause {
-                    span: Span::default(),
-                    xtor: "Cons".to_owned(),
-                    context: ctx_case,
-                    rhs: Var {
-                        span: Span::default(),
-                        var: "x".to_owned(),
-                        ty: Some(Ty::mk_i64()),
-                    }
-                    .into(),
-                },
-            ],
-            destructee: Rc::new(
-                Var {
-                    span: Span::default(),
-                    var: "x".to_owned(),
-                    ty: Some(Ty::mk_decl("ListInt")),
-                }
-                .into(),
-            ),
-            ty: Some(Ty::mk_i64()),
-        };
-        assert_eq!(result, expected)
-    }
-
-    #[test]
-    fn check_case_tup() {
-        let mut ctx_case = TypingContext::default();
-        ctx_case.add_var("x", Ty::mk_i64());
-        ctx_case.add_var("y", Ty::mk_i64());
-        let mut ctx = TypingContext::default();
-        ctx.add_var("x", Ty::mk_decl("TupIntInt"));
-        let symbol_table = symbol_table_tup();
-        let result = Case {
-            span: Span::default(),
-            cases: vec![Clause {
-                span: Span::default(),
-                xtor: "Tup".to_owned(),
-                context: ctx_case.clone(),
-                rhs: Var::mk("x").into(),
-            }],
-            destructee: Rc::new(Var::mk("x").into()),
-            ty: None,
-        }
-        .check(&symbol_table, &ctx, &Ty::mk_i64())
-        .unwrap();
-        let expected = Case {
-            span: Span::default(),
-            cases: vec![Clause {
-                span: Span::default(),
-                xtor: "Tup".to_owned(),
-                context: ctx_case,
-                rhs: Var {
-                    span: Span::default(),
-                    var: "x".to_owned(),
-                    ty: Some(Ty::mk_i64()),
-                }
-                .into(),
-            }],
-            destructee: Rc::new(
-                Var {
-                    span: Span::default(),
-                    var: "x".to_owned(),
-                    ty: Some(Ty::mk_decl("TupIntInt")),
-                }
-                .into(),
-            ),
-            ty: Some(Ty::mk_i64()),
-        };
-        assert_eq!(result, expected)
-    }
-
-    #[test]
-    fn check_case_fail() {
-        let mut ctx = TypingContext::default();
-        ctx.add_var("x", Ty::mk_i64());
-        ctx.add_var("y", Ty::mk_i64());
-        let symbol_table = symbol_table_list();
-        let result = Case {
-            span: Span::default(),
-            cases: vec![Clause {
-                span: Span::default(),
-                xtor: "Tup".to_owned(),
-                context: ctx,
-                rhs: Var::mk("x").into(),
-            }],
-            destructee: Rc::new(Lit::mk(1).into()),
-            ty: None,
-        }
-        .check(&symbol_table, &TypingContext::default(), &Ty::mk_i64());
-        assert!(result.is_err())
-    }
-
-    fn example_empty() -> Case {
-        Case {
-            span: Span::default(),
-            destructee: Rc::new(Var::mk("x").into()),
-            cases: vec![],
-            ty: None,
-        }
-    }
-
-    fn example_tup() -> Case {
-        let mut ctx = TypingContext::default();
-        ctx.add_var("x", Ty::mk_i64());
-        ctx.add_var("y", Ty::mk_i64());
-        Case {
-            span: Span::default(),
-            destructee: Rc::new(Var::mk("x").into()),
-            cases: vec![Clause {
-                span: Span::default(),
-                xtor: "Tup".to_owned(),
-                context: ctx,
-                rhs: Term::Lit(Lit::mk(2)),
-            }],
-            ty: None,
-        }
-    }
-
-    #[test]
-    fn display_empty() {
-        assert_eq!(
-            example_empty().print_to_string(Default::default()),
-            "x.case { }"
-        )
-    }
-
-    #[test]
-    fn parse_empty() {
-        let parser = fun::TermParser::new();
-        assert_eq!(parser.parse("x.case { }"), Ok(example_empty().into()));
-    }
-
-    #[test]
-    fn display_tup() {
-        assert_eq!(
-            example_tup().print_to_string(Default::default()),
-            "x.case { Tup(x: i64, y: i64) => 2 }"
-        )
-    }
-
-    #[test]
-    fn parse_tup() {
-        let parser = fun::TermParser::new();
-        assert_eq!(
-            parser.parse("x.case { Tup(x : i64, y : i64) => 2 }"),
-            Ok(example_tup().into())
-        );
-    }
-}
